pyproject.toml
[tool.poetry]
name = "triviabot"
version = "0.1.1"
description = ""
authors = ["Zenith <z@zenith.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.4.1"
"discord.py-stubs" = "^1.4.0"
google-api-python-client = "^1.11.0"
aurcore = "^0.1.29"
aurflux = "^1.2.71"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.1"
gspread = "^3.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
